SSCSSI-439: removing abstract ITs
